ES: Performance improvement when invalidating submissions.
Use a joined load to avoid having to fetch submission results all the time.
Time taken to invalidate 1200 submission drops from >30s to about 10s.
This could be improved further still, possibly by issuing a mass delete and
relying on the database cascading rules to remove evaluations and executables.
But that sounds a little bit too scary for now.